+++ ref/mir_scheduler/run_job_script.py
@@ -1,0 +1,7 @@
+import sys
+sys.path.append('../mir/')
+
+from app.conf import const
+from app.util import util_panel, util_db, util_agent, util_library
+
+print(const.CONF)